# Lazier ticks for large speedup when creating large arrays of axes.